+++ changelog.md
@@ -2,6 +2,7 @@ little-saga 在一些功能上会和 redux-saga 保持同步，🔗 表示在 re
 
 ## (unreleased)
 
+- 新增 GET_ENV effect
 - 导出 makeMatcher 函数
 
 ## v0.5.1